% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as_tensor.array}
\alias{as_tensor.array}
\title{Convert an Array to a Tensor}
\usage{
\method{as_tensor}{array}(array, device)
}
\arguments{
\item{array}{An array to be converted to a tensor.}

\item{device}{A CUDA device object to associate with the tensor.}
}
\value{
A tensor object.
}
\description{
Converts an n-dimensional array to a tensor.
}
