% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as_tensor.matrix}
\alias{as_tensor.matrix}
\title{Convert a Matrix to a Tensor}
\usage{
\method{as_tensor}{matrix}(array, device)
}
\arguments{
\item{array}{A matrix to be converted to a tensor.}

\item{device}{A CUDA device object to associate with the tensor.}
}
\value{
A tensor object.
}
\description{
Converts a matrix to a tensor. Note: This performs an implicit transpose of the matrix.
}
