% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{as_tensor}
\alias{as_tensor}
\title{Convert an Object to a Tensor}
\usage{
as_tensor(array, device)
}
\arguments{
\item{array}{The input data to be converted to a tensor.}

\item{device}{A CUDA device object to associate with the tensor.}
}
\value{
A tensor object.
}
\description{
Generic function to convert different data types to a tensor.
}
