% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{collect_tensor}
\alias{collect_tensor}
\title{Collect Tensor Data}
\usage{
collect_tensor(tensor)
}
\arguments{
\item{tensor}{A tensor object to be collected.}
}
\value{
An R array, matrix, or vector depending on the tensor's dimensions.
}
\description{
Converts the contents of a tensor back into an R array or matrix.
}
