% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{new_cuda}
\alias{new_cuda}
\title{Create a New CUDA Device}
\usage{
new_cuda(ordinal)
}
\arguments{
\item{ordinal}{An integer specifying the ordinal index of the CUDA device.}
}
\value{
A CUDA device object.
}
\description{
Initializes and returns a new CUDA device object for GPU programming.
}
