% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{new_tensor}
\alias{new_tensor}
\title{Create a New Tensor}
\usage{
new_tensor(array, dims, device)
}
\arguments{
\item{array}{A numeric vector or array to initialize the tensor.}

\item{dims}{An integer vector specifying the dimensions of the tensor.}

\item{device}{A CUDA device object to associate with the tensor.}
}
\value{
A tensor object.
}
\description{
Constructs a tensor object using the specified array, dimensions, and device.
}
