% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{+.Tensor}
\alias{+.Tensor}
\title{Add Two Tensors}
\usage{
\method{+}{Tensor}(lhs, rhs)
}
\arguments{
\item{lhs}{The left-hand side tensor.}

\item{rhs}{The right-hand side tensor.}
}
\value{
A tensor resulting from the addition.
}
\description{
Performs element-wise addition of two tensors.
}
