% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{print.Tensor}
\alias{print.Tensor}
\title{Print a Tensor Object}
\usage{
\method{print}{Tensor}(tensor, ...)
}
\arguments{
\item{tensor}{A tensor object to be printed.}

\item{...}{Additional arguments passed to the printing method.}
}
\description{
Prints a human-readable representation of a tensor object.
}
